begonia: Sync extractors with templates
One can notice missing device name from header comments in vendor's
makefiles, blueprint & soong namespace. This is happening because
we are passing 'true' as 4th arg in setup_vendor function.

While just removing that fixes the issue, a better solution is to synce the
scripts with the templates altogther. This fixes copyrights, formatting,
as well as gets rid of parameters which should be passed only on common devices.

Signed-off-by: Aayush Gupta <aayushgupta219@gmail.com>